Move organization-service onto loco-rs 1.0.1
Second crate on the family-wide loco-rs 1.0.1 migration, using the
recipe proven on authentication-service (the reference crate). Same
mechanical fixes: sea-orm's raw-Statement calls in review_queue.rs move
to execute_raw/query_one_raw/query_all_raw; a stray .into() in
event_outbox.rs became a useless_conversion once Expr::current_timestamp
already returns the right type; and loco's ColType::PkAuto now generates
a 64-bit primary key, so the organizations/audit_logs/merge_records
entities (and the compliance code carrying their row ids) move from i32
to i64.

One thing worth carrying into the rest of the migration: event_outbox's
id column deliberately stays i32/SERIAL, because its migration writes
raw SQL rather than going through the loco schema DSL (to control the
exact un-pluralized table name) — so it never picked up the DSL's width
change. Any crate with a hand-rolled-SQL table needs checking against
its actual column type, not just against whether ColType::PkAuto was
used elsewhere in the same crate.

Verified: cargo build --all-targets, fmt, clippy -D warnings, cargo test
(137 passed) clean; the full DB-gated suite (26 tests) green on the
first run against a freshly migrated Postgres 18 — the i32→i64 widening
was done proactively this time rather than discovered via a runtime
UnpackInsertId error.

+++ organization/organization-service-with-loco/src/compliance/record_integrity.rs
@@ -322,7 +322,7 @@ mod tests {
     }
 
     /// A stored row with a correct hash.
-    fn row(id: i32, name: &str) -> organizations::Model {
+    fn row(id: i64, name: &str) -> organizations::Model {
         let mut model = organizations::Model {
             created_at: at(1_700_000_000_000_000),
             updated_at: at(1_700_000_000_000_000),
